ACTIN-46: split out separate algo class
